joybridge_host/pyproject.toml
[project]
name = "joybridge-host"
version = "0.1.0"
description = "Python interface for using a USB dongle as an ESP-NOW transponder"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1.8",
    "ipdb>=0.13.13",
    "pygame>=2.6.1",
    "pyserial>=3.5",
]
